xls reader file
This is an xls_reader file compatible with apache poi 4.1.0.
just ignore the package name and use yours in place of mine but create a class under YOUR java project file structures on the package you have called xls_Reader in this case in selenium IDE and then copy the rest of the code in this uploaded file as it is. Ensure to have referenced and imported all the poi 4.1.0 jar files.